gasification/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-filter==24.3
django-prometheus==2.3.1
django-rest-passwordreset==1.4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
factory_boy==3.3.1
Faker==33.1.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
prometheus_client==0.21.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
